Datamapper update
git-svn-id: https://svn.eu.apache.org/repos/asf/incubator/roller/branches/roller_3.0@443151 13f79535-47bb-0310-9956-ffa450edef68

+++ sandbox/jdobackend/src/org/apache/roller/business/datamapper/DatamapperPingQueueManagerImpl.java
@@ -43,11 +43,11 @@ public class DatamapperPingQueueManagerImpl implements PingQueueManager {
     public void addQueueEntry(AutoPingData autoPing) 
             throws RollerException {
         // first, determine if an entry already exists
-        int count = (Integer)strategy.newQuery(PingQueueEntryData.class,
+        Integer count = (Integer)strategy.newQuery(PingQueueEntryData.class,
                 "countGetByPingTarget&&website")
                 .execute(new Object[]
                     {autoPing.getPingTarget(), autoPing.getWebsite()});
-        if (count > 0)
+        if (count.intValue() > 0)
             return;
 
         // create and store a new entry
